requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0

# Configuration and Utilities
hydra-core>=1.3.0
omegaconf>=2.3.0
PyYAML>=6.0

# Data Processing and Analysis
pandas>=2.0.0
numpy<2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
seaborn>=0.12.0

# Image Processing and Augmentation
Pillow>=9.5.0
albumentations>=1.3.0
imgaug>=0.4.0
opencv-python>=4.7.0

# Experiment Tracking and Logging
wandb>=0.15.0

# Additional Utilities
tqdm>=4.65.0
pyyaml>=6.0
seaborn

# Quantization
brevitas>=0.10.0

# ONNX export & numerical validation
onnx>=1.14.0
onnxruntime>=1.16.0
onnxscript
onnxoptimizer
qonnx

# Development (Optional)
# jupyter>=1.0.0
# tensorboard>=2.13.0
